python_src/requirements.txt
antlr4-python3-runtime==4.9.2
coverage==6.1.1
pyeda==0.28.0
PySMT==0.9.0
six==1.16.0
z3-solver==4.8.12.0
